Public API: Replace Health() by SupportsTarget() (#18)
While integrating go-libddwaf v1.3.0 into dd-trace-go, we figured out
that `Load()` was doing too much for the case of remote activation of
appsec when in a disabled state, where we don't want to load libddwaf at
all until appsec actually starts up remote activation order.

For this reason, we decided to make `Health()` useful again by replacing
it with `SupportsTarget()` in order to be able to only check if we are
on a supported target or not. This allows us to properly check for
proper support of libddwaf, apart from the libddwaf loading, to do it
early in the appsec startup and be able to properly log the information
to the user.