feat: improved WAF init error reporting
The `ddwaf_init` call attempts to provide more information about failure
causes through the `diagnostics` object returned. This may be populated
even when receiving a `NULL` handle.

This PR attempts to obtain an error from diagnostics when available in
order to provide improved error messaging to the callers.

Additionally, if diagnostics include any "top-level" error (via the
`DiagnosticEntry.Error` field), consider the WAF initialization as failed
and return the specified error. Errors reported at the individual entity
level are not considered in this new check, as they may stem from new
rules using operators that are not yet available in the current WAF
library version, and that should be okay.

This saves all users of the WAF from having to manually check for these
error conditions in all places, which is cumbersome and error-prone.