feat: WAFObject encoder produces truncation information
Instead of "silently" performing truncations unless they are required on
the top-level value, the encode now keeps tabs on the truncations it
performed, recording the set of un-truncated sizes for each truncation
cause.

This information will subsequently be made visible to downstream WAF
users so that it is possible to implement telemetry tracking of the
real-world impact of encoding limits, and possibly gain insight into how
to tune them for maximum efficiency.