Ensure WAF static libraries get tracked in vendor/
This ensures that WAF static libraries are not ignored from source control by parent .gitignore files. This situation can arise when a project uses Go vendoring and depends on this library. Because it is common for Go projects to gitignore all so/dylib files, this vendored dependency will not prompt to track these files. When the project is checked out, the vendored dependency will be incomplete (missing the 2 .so and 2 .dylib files).