fix: fix conflicting symbols between v2 & v3 (#100)
Go semantic versionning makes it that we can have multiple major
versions of the same module in the same binary. Since we exported an
hard-coded symbol it breaks this possibility. This should not happen in
practice for customers because noone is using go-libddwaf as-is except
dd-trace-go & some DataDog backend code

The error:
```
/usr/local/go/pkg/tool/linux_amd64/link: running gcc failed: exit status 1
/bin/ld: /tmp/go-link-2237125540/000030.o: in function `ddwafLogCallbackFn':
/tmp/go-build/_cgo_export.c:27: multiple definition of `ddwafLogCallbackFn'; /tmp/go-link-2237125540/000023.o:/tmp/go-build/_cgo_export.c:27: first defined here
```

Signed-off-by: Eliott Bouhana <eliott.bouhana@datadoghq.com>