fix: libddwaf does not accept nullptr-valued blank strings (#120)
Inadvertently replaced the pointer to a `\x00` byte that was used when
encoding a blank string with a `0` (`null`) pointer; which `libddwaf`
does not currently accept.

This commit restores the previous encoding behavior, and adds a test
that was created based on the ruleset that allowed discovery of this
issue.

---

Also fixes up the CGO-exported symbol name to be at V4 so it does not
clash with the one exported by V3.