chore: upgrade to libddwaf 1.25.1 (#128)
New API surface: the `Result` object now has a `Keep` field that is used
by the WAF to indicate a trace should be set to manual keep priority.

---

Other changes prompted by this:

- The `WAFObject` struct now carries decoding methods that are more
convenient to use & easier to discover than their previous free-floating
function incarnation;
- The `unsafe.CString` function now accepts a `*runtime.Pinner` as it
allocates a slice which is pointed to by the result; but this is often
incorrectly handled resulting in a possible GC race;
- Stopped calling `runtime.KeepAlive` on values that are actually not on
the heap (i.e, `uintptrs` originating from C-side which are passed
by-value everywhere), as these were strictly unnecessary;
- Using `runtime.Pinner` in a bunch more places where pointers were
previously just kept alive via `runtime.KeepAlive`; but which are passed
to C and hence need to be pinned, too.