fix: data race between `libddwaf.Load()` and `libddwaf.Usable()`
A data race existed between `libddwaf.Load()` and `libddwaf.Usable()` as
both access shared global state that is not managed through atomic pointers.

This PR puts the shared state behind atomics so that the data race no longer
exists. We copy the `wafLib` pointer in the `Builder`, `Handle` and
`Context` structures to avoid having to load the atomic all the time,
which would have broad performance impact.